Final student code for processor class

+++ include/processor.h
@@ -7,7 +7,7 @@ class Processor {
   Processor();
 
 // Memeber Function
-  float Utilization(); 
+  double Utilization(); 
 
   // Declare any necessary private members
  private:
